requirements.txt
# Web Framework
flask==3.0.0
gunicorn==21.2.0
flask-cors==4.0.0

# OpenAI
openai==1.3.5

# HTTP Client
requests==2.31.0

# Utilities
python-dotenv==1.0.0
sympy==1.12
beautifulsoup4==4.12.2

# Telegram Bot
python-telegram-bot==20.6

# Remove transformers if not actively using it
# transformers==4.35.2